[cgroups2] Add allow / deny list normalization validation.
Currently we assume that a device state is normalized before using it
for generating ebpf files. However, we have not been enforcing these
constraints on the device access state.

We enforce some basic validation on cgroups2::configure on the state
to ensure that we are able to generate a correct ebpf program. If the
lists are not normalized, we generate incorrect programs!

An allow or deny list is 'normalized' iff everything below are true:

  1. No entries have empty accesses specified.
  2. No two entries on the same list can have the same selector
     (type, major & minor numbers).
  3. No two entries on the same list can be encompassed by the other
     entry. See Entry::encompassed.

This patch adds helpers to check if a device state is normalized,
and will only allow users to create new CgroupDeviceAccess instances
using a helper that checks that the allow and deny lists are normalized.

A new helper function is added to check if an entry would be granted
access, and requires the state to be normalized.

Review: https://reviews.apache.org/r/75099/