[cgroups2] Helper to check device entry normalization.
Currently we assume that a device state is normalized before using it
for generating ebpf files. However, we have not been enforcing these
constraints.

We add a helper to check if a device state is normalized so that we can
enforce these constraints.

An allow or deny list is 'normalized' iff everything below are true:

1. No Entry has empty accesses specified.
2. No two entries on the same list can have the same selector (type,
   major & minor numbers).
3. No two entries on the same list can be encompassed by the other
   entry (see Entry::encompasses).

Review: https://reviews.apache.org/r/75099/