[cgroups2] Helper to check device access.
A device access is granted if it is encompassed by an allow entry and
does not have access overlaps with any deny entry.

The current process of manually checking if a device access would be
granted given a state is tedious and leads to worse readability.

A new helper function is added to check if an entry would be granted
access in a CgroupDeviceAccess instance, and requires the state to be
normalized.

Review: https://reviews.apache.org/r/75113/