[cgroups2] Enforce normalization in configure.
We currently do not enforce normalized allow and deny in configure.
However, to ensure that we can generate an ebpf program that behaves
correctly, we have to ensure that allow and deny are normalized.

This patch adds a validation check to ensure that the allow and deny are
normalized before attempting to generate the ebpf program.

Review: https://reviews.apache.org/r/75114/