[devices] Enforce normalization for DeviceManager configure & reconfigure.
Currently in the configure() and reconfigure() functions in device
manager, we do not ensure that the device access state at the end of the
function call is normalized. So we incorporate normalized() and
normalize() calls to ensure that the allow and deny lists are always
normalized at the end of a configure() or reconfigure() call.

Review: https://reviews.apache.org/r/75115/