[cgroups2] Support DeviceManager in GPU isolator.
Currently, the GPU isolator assumes we are only using cgroups v1, and
makes use of the cgroups::devices::allow and deny functions to control
GPU access.

In Cgroups2, we need to attach ebpf programs for the specific cgroups,
which is done for us in the DeviceManager. Hence, we need to use the
DeviceManager in the GPU isolator depending on whether cgroups v1 or v2
is currently mounted.

Review: https://reviews.apache.org/r/75074/