[device manager] Checkpoint state on device manager state change.
Currently we do not checkpoint the device access state of each cgroup
when the configure or reconfigure is called. Meaning that we have
no way of recovering a cgroup's device access state.

We will checkpoint state of the device manager whenever its state
is being changed to ensure that we can recover the most recent state
when necessary.

Review: https://reviews.apache.org/r/75143/