[cgroups2] Prevent containerId from prepending cgroup root during recovery.
Currently, CgroupsIsolatorTest.ROOT_CGROUPS_PERF_PerfForward fails
during recovery because it cannot create the directory for the recovered
container. This happens because the original containerId, when
recovered, includes the cgroup root. The function that converts a cgroup
to a containerId does not ignore the cgroup root, even though we do not
expect it to be included.

To fix this, we will remove the first token of the cgroup if it matches
the cgroup root. This will prevent attaching an extraneous cgroup root
to the containerId when parsing it from a cgroup.

Review: https://reviews.apache.org/r/75156/