[cgroups2] Collect process & thread from leaf groups.
Currently, our core controller usage() function try to read the cgroup
files in the argument cgroup. However, in our design for cgroups v2,
processes and threads live in the leaf child of a cgroup. Hence the
usage collection will not find the actual processes and threads for a
cgroup if it's not already specified as a leaf group.

This patch adds a check to see if the argument cgroup is a leaf cgroup,
and will search for processes and threads in the leaf cgroup instead.

Review: https://reviews.apache.org/r/75158/