[cgroups2] Create memory controller using cgroups/all flag.
Currently, we cannot use the cgroups/all flag to its corresponding
controllers in cgroups2.

The flag causes us to grabs all the values in the cgroup.controllers
file. But we should instead just add all the creators when we see
cgroups/all, as many controllers are no longer present in
cgroup.controllers in cgroups v2.

This fix incidentally fixes the
ROOT_CGROUPS_AgentRecoveryWithNewCgroupSubsystems test as it was unable
to create the memory controller using cgroups/all flag.

Review: https://reviews.apache.org/r/75162/