[cgroups2] Fix ROOT_CGROUPS_AutoLoadSubsystems test.
Currently, the ROOT_CGROUPS_AutoLoadSubsystems test is failing because
it is checking for hierarchies for subsystems enabled under
'cgroups/all'. In cgroups2 we cannot perform this check because of the
unified hierarchy.

Hence we skip this hierarchy check and instead check that all available
cgroups2 controllers are enabled by reading cgroup.controllers and
cgroup.subtree_control.

Review: https://reviews.apache.org/r/75163/