[cgroups2] Add device manager recovery support.
We currently do not have any method of recovering the device access
states when the cgroups2 isolator is atempting to recover containers.

We add a recovery state here that makes use of the protobuf checkpoint
files to ensure that the previous device accesses of cgroups can be
restored. It will be used by the cgroups2 isolator.

Review: https://reviews.apache.org/r/75145/