[cgroups2] Enable controller in parent cgroups during prepare().
To support nested containers with nested cgroups, we need to enable
controllers in cgroup.subtree_control file for the appropriate nested
cgroup.

To do so, we need to ensure that the parents have the the requested
controller in their cgroup.subtree_control file so that the nested
cgroup can have the controller written into subtree_control as well.
Otherwise we will get a 'no such file or directory' error.

Review: https://reviews.apache.org/r/75166/