[cgroups2] Make isolator update() fail for containers that don't need isolation.
In cgroups1, we returned an error when we see that the container is
sharing cgroups, as it would have no cgroups for update() to take.

In cgroups2 we will mimick this behavior for containers that do not
wish to have their own resource constraints, as we do not expect to call
update() on those containers.

Review: https://reviews.apache.org/r/75174/