[cgroups2] Make isolator status() return parent status for containers w/ !isolate.
In cgroups1, if a container is sharing cgroups with its parents, we will
return the parent's status.

In cgroups2, we want to mimic this behavior even though we always create
cgroups for our containers. Since we do not do anything with our
container's cgroup if isolate == false.

Review: https://reviews.apache.org/r/75175/