[cgroups2] Handle unknown containers in watch().
In cgroups1, we returned a pending future for containers that shared
cgroups with their parents.

In cgroups2, since we always create cgroups for our containers, we no
longer need to consider this special case. So we only return failure
if there is an unknown container.

Review: https://reviews.apache.org/r/75176/