[cgroups2] cgroups2::destroy retry rmdir on EBUSY.
Currently we only wait until a cgroup's pids (retrieved from
cgroup.procs) is empty. However even if a cgroup's pids are empty the
rmdir call on it may still return EBUSY, causing us to fail the destroy
operation. We want to retry the rmdir operation even on EBUSY for up to
5 seconds to ensure that we are able to delete the cgroup.

This approach is similar to how crun is destroying its cgroups.
see:

https://github.com/containers/crun/blob/10b3038c1398b7db20b1826f94e9d4cb444e9568/src/libcrun/cgroup-utils.c#L471

Review: https://reviews.apache.org/r/75181/