[cgroups2] Remove completed todos.
The Todos mention recovery support for the device manager, and to use
the device manager recover in the device controller process.

We have added checkpointing and recovery support in:
https://reviews.apache.org/r/75145

We also only use the device manager recovery in the cgroups2 isolator
instead of in the device controller, as implemented in:
https://reviews.apache.org/r/75149/

As such, these todos are considered completed, and can be removed.

Review: https://reviews.apache.org/r/75189/